% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs_utils.R
\name{assert_projected}
\alias{assert_projected}
\title{Assert That an sf Object is in a Projected CRS}
\usage{
assert_projected(x, arg_name = "x")
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{sfg} object.}

\item{arg_name}{Character scalar. Name of the argument as it appears in the
calling function, used in the error message. Default \code{"x"}.}
}
\value{
Invisibly returns \code{TRUE} if the CRS is projected. Otherwise
throws an error.
}
\description{
Checks that \code{x} is in a projected (non-geographic) CRS and throws an
informative error if not. This is the internal CRS guard used at the top of
all \pkg{spatintegrate} geometry functions.
}
\examples{
proj <- sf::st_sfc(sf::st_point(c(0, 0)), crs = 32619)
assert_projected(proj)  # silent

\dontrun{
geo <- sf::st_sfc(sf::st_point(c(-71, 42)), crs = 4326)
assert_projected(geo)  # error
}

}
\seealso{
\code{\link[=ensure_projected]{ensure_projected()}}, \code{\link[=ensure_same_crs]{ensure_same_crs()}}
}
