% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_prior.R
\name{fourier_phi_to_sar_phi}
\alias{fourier_phi_to_sar_phi}
\title{Convert Fourier phi back to SAR phi}
\usage{
fourier_phi_to_sar_phi(phi_fourier, rho, delta, freq_grid)
}
\arguments{
\item{phi_fourier}{Positive numeric. phi from Fourier model.}

\item{rho}{Numeric in (0,1). SAR spatial autoregression parameter.}

\item{delta}{Positive numeric. Pixel size in CRS units.}
}
\value{
Positive numeric. Equivalent SAR phi.
}
\description{
Inverse of sar_phi_to_fourier_phi().
}
