% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_matern_matching.R
\name{physical_to_unit_square}
\alias{physical_to_unit_square}
\title{Convert physical Matern parameters to unit-square TAU and kappa2_norm}
\usage{
physical_to_unit_square(kappa2_phys, sigma2_M_scale, domain_bbox)
}
\arguments{
\item{kappa2_phys}{Physical kappa^2 from sar_to_matern().}

\item{sigma2_M_scale}{Matern scale from sar_to_matern().}

\item{domain_bbox}{Length-4 numeric (xmin,ymin,xmax,ymax).}
}
\value{
Named list: kappa2_norm, TAU, L, Lx, Ly.
}
\description{
Given physical kappa2 and sigma2_M_scale from sar_to_matern(), and the
domain bounding box, returns the unit-square parameters for fourier_matern_Q_sb().
}
\details{
Relationships:
kappa2_norm = kappa2_phys * L^2        (L = max(Lx,Ly))
TAU         = 1 / (sigma2_M_scale * Lx * Ly)
}
