% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_prior.R
\name{sar_phi_to_fourier_phi}
\alias{sar_phi_to_fourier_phi}
\title{Convert SAR phi to Fourier phi (physical units)}
\usage{
sar_phi_to_fourier_phi(phi_sar, rho, delta, freq_grid)
}
\arguments{
\item{phi_sar}{Positive numeric. phi from SAR CV fit.}

\item{rho}{Numeric in (0,1).}

\item{delta}{Positive numeric. Pixel size in CRS units.}

\item{freq_grid}{List from fourier_freq_grid().}
}
\value{
Positive numeric. phi for the Fourier model.
}
\description{
phi_fourier = phi_sar * 16 * Lx * Ly / (rho^2 * delta^2)
}
