% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sar_matern_matching.R
\name{unit_square_to_physical}
\alias{unit_square_to_physical}
\title{Convert unit-square parameters to physical Matern parameters}
\usage{
unit_square_to_physical(kappa2_norm, TAU, domain_bbox)
}
\arguments{
\item{kappa2_norm}{Unit-square kappa^2.}

\item{TAU}{Unit-square precision.}

\item{domain_bbox}{Length-4 numeric (xmin,ymin,xmax,ymax).}
}
\value{
Named list: kappa2_phys, sigma2_M_scale.
}
\description{
Inverse of physical_to_unit_square().
}
