% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_prior.R
\name{fourier_phi_to_sar_phi}
\alias{fourier_phi_to_sar_phi}
\title{Convert Fourier phi back to SAR phi}
\usage{
fourier_phi_to_sar_phi(phi_fourier, rho, delta, freq_grid)
}
\description{
Convert Fourier phi back to SAR phi
}
