% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qmc_utils.R
\name{generate_qmc_unit_square}
\alias{generate_qmc_unit_square}
\title{Generate Sobol QMC Points in the Unit Square}
\usage{
generate_qmc_unit_square(n)
}
\arguments{
\item{n}{Positive integer. Number of QMC points to generate.}
}
\value{
An \eqn{n \times 2} numeric matrix. All values are in \eqn{[0, 1]}.
}
\description{
Returns an \eqn{n \times 2} matrix of quasi-Monte Carlo points in
\eqn{[0,1]^2} using a scrambled Sobol sequence. Points are suitable for
mapping into triangle or quadrilateral regions via
\code{\link[=map_unit_square_to_triangle]{map_unit_square_to_triangle()}} or \code{\link[=map_unit_square_to_quad]{map_unit_square_to_quad()}}.
}
\examples{
pts <- generate_qmc_unit_square(128)
stopifnot(all(pts >= 0), all(pts <= 1))

}
